refactor: move to static method on treeContainer
